winex11.drv: Implement cloaking using window opacity and Xshape.
Cloaking is implemented by setting the opacity of a window to zero, and
setting an empty shape using libXshape. This makes it transparent to mouse
input and completely invisible, while otherwise still processing the exact
same messages/events as a visible window would, to match what Windows does.

While a cloaked window still has a taskbar visible, it also has the same
traits as a normal window. For example, a minimized cloaked window will show
the "Restore" or "Unminimize" function when right-clicking on its taskbar
icon. So that means unmapping the window would *not* be the same behavior
as Windows, since it even removes the taskbar item in the first place.

Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>