winex11.drv: Implement WM virtual desktops via shell cloaking.
When a window is on a virtual desktop, the WM iconifies it without marking
it "hidden". We use that to mark it as "shell cloaked" instead of minimizing
the window, to match what Windows does when switching virtual desktops. This
fixes applications which do things when minimized, since they are *not*
minimized on Windows when switching desktops, unlike on Wine currently.

When on a different desktop, GetWindowCompositionAttribute always reports
DWM_CLOAKED_SHELL. The shell cloaking the window (when on a different desktop)
is tracked independently of the manual cloaking done by an app. It always
cloaks it when on a different desktop, even if the app asks the window to
be uncloaked. However, when going back to the original desktop, the window
will be uncloaked (because the app uncloaked it while off-desktop). So
even if DWM_CLOAKED_SHELL is reported, it's possible that the app is still
manually cloaked underneath, with no way to know until you switch back to
the desktop containing it and calling the API again.

Windows on a different desktop (i.e. cloaked by the shell) can't be set to the
foreground. It simply fails. If some item on the desktop is selected and has
focus, then switching desktops keeps it selected and no window gets activated
on a desktop switch, because the desktop items are common between desktops.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=43691
Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>