ntdll: Convert Windows Epoch to UNIX Epoch
Some games are hanging on startup, with futexes waiting forever (even
with timeout values). This is because the timeout values are been passed
as Windows Epoch, instead of UNIX one. Convert the values so timeouts
work again. Also, use realtime clock instead of monotonic one, so we
don't need to issue another syscall do to this conversion.