wineboot: Return TSC frequency in ~Mhz
Some games such as Horizon Zero Dawn use this registry value to correlate values from rtdsc -> real time.

In my testing across a few devices, Windows always returns the tsc frequency in this entry, not the current/maximum frequency of the processor.

Returning the nominal/maximum cpu frequency here causes the game to run in slow motion as it may not match the tsc frequency of the processor.

Ideally we'd not have to measure this and the kernel would return tsc_khz to userspace, but this is a good enough stop-gap until https://lkml.org/lkml/2020/12/31/72 or something similar is merged.

Fixes: #4125 (Slow motion bug)

Signed-off-by: Joshua Ashton <joshua@froggi.es>